feat!: migrate to burn 0.21 and replace burn-ndarray with burn-flex
burn 0.21.0 shipped 2026-05-07 and upstream deprecated burn-ndarray in
the 0.22 cycle, replacing it with burn-flex. This moves the CPU delegate
to burn-flex now rather than after burn-ndarray is removed.

Backend trait changes:
  - split the associated types out into BackendTypes; the three
    near-identical Backend impls collapse into one, since only the
    reported name differed
  - supports_dtype -> dtype_usage returning DTypeUsageSet
  - Device::device_count moved onto Backend

Op signature changes:
  - tensor-creating and comparison ops take an explicit output dtype
    (bool_empty/zeros/ones, *_into_int, *_into_float, all comparisons,
    argmax/argmin, random, full, dequantize, q_argmax/q_argmin)
  - scalar params are now the Scalar enum rather than raw f32/i64/bool;
    values crossing the C FFI convert with .elem::<f32>()
  - Shape fields are private: shape.dims -> .to_vec()/.num_elements()
  - new required ops implemented: float_argtopk, int_argtopk,
    attention, rfft, irfft
  - int_powf and int_powf_scalar_impl were removed from IntTensorOps

BREAKING: IntElem is now i32 rather than i64, matching burn-flex and
burn's ecosystem default, so delegated int ops need no dtype conversion
per call. Tests reading argmax/int output as Vec<i64> now read Vec<i32>.

Conversion helpers are substantially simpler: burn-flex exposes a single
FlexTensor with runtime dtype dispatch and TensorData constructors, so
the raw ndarray shape/layout juggling is gone, along with the direct
ndarray dependency. The four hand-inverted comparison bodies in the
apple path (greater_equal = NOT less, lower_equal = NOT greater) now
share one float_handle_to_inverted_bool helper.

Also refreshes dependencies left behind since April: openvino 0.7->0.11,
candle 0.8->0.11, safetensors 0.4->0.8, tokenizers 0.21->0.23, dirs
5->6, indicatif 0.17->0.18. Declares rust-version = 1.89 (burn 0.21's
MSRV). The bench now reports burn-flex as the CPU baseline.

Verified: fmt, clippy -D warnings, and cargo check --all-targets across
default/apple/intel/qualcomm; 31 tests pass on apple and default.

+++ tests/burn_backend.rs
@@ -177,7 +177,7 @@ fn max_min() {
 #[test]
 fn argmax_dim() {
     let a = Tensor::<B, 2>::from_floats([[1.0, 3.0, 2.0], [5.0, 4.0, 6.0]], &dev());
-    let idx: Vec<i64> = a.argmax(1).into_data().to_vec().unwrap();
+    let idx: Vec<i32> = a.argmax(1).into_data().to_vec().unwrap();
     assert_eq!(idx, vec![1, 2]);
 }
 
@@ -274,7 +274,7 @@ fn zeros_and_ones() {
 #[test]
 fn int_from_data() {
     let a = Tensor::<B, 1, burn::tensor::Int>::from_ints([1, 2, 3], &dev());
-    let d: Vec<i64> = a.into_data().to_vec().unwrap();
+    let d: Vec<i32> = a.into_data().to_vec().unwrap();
     assert_eq!(d, vec![1, 2, 3]);
 }
